// packages/client/src/store/leaderboard/enum.ts
export enum LeaderboardEnum {
  InitFetchLeaderboard = 'INIT_FETCH_LEADERBOARD',
  SuccessFetchLeaderboard = 'SUCCESS_FETCH_LEADERBOARD',
  FailFetchLeaderboard = 'FAIL_FETCH_LEADERBOARD',
  InitFetchUserRank = 'INIT_FETCH_USER_RANK',
  SuccessFetchUserRank = 'SUCCESS_FETCH_USER_RANK',
  FailFetchUserRank = 'FAIL_FETCH_USER_RANK',
  SetSelectedPeriod = 'SET_SELECTED_PERIOD',
  CleanLeaderboardAlerts = 'CLEAN_LEADERBOARD_ALERTS',
}